new MyApp.dart . starting to implement modular part
